\name{assessment-package}
\alias{assessment-package}
\alias{assessment}
\docType{package}
\title{
\packageTitle{assessment}
}
\description{
\packageDescription{assessment}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{assessment}
\packageIndices{assessment}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{assessment}

Maintainer: \packageMaintainer{assessment}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
