% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{get_assessmemt}
\alias{get_assessmemt}
\title{Get assessment data
Expects assessment.dbname, assessment.host, assessment.user, assessment.password to be set for access to the assessment database}
\usage{
get_assessmemt(geography = NA)
}
\description{
Get assessment data
Expects assessment.dbname, assessment.host, assessment.user, assessment.password to be set for access to the assessment database
}
