% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{get_buildings}
\alias{get_buildings}
\title{Get buildings data
Expects assessment.dbname, assessment.host, assessment.user, assessment.password to be set for access to the assessment database}
\usage{
get_buildings(geography = NA)
}
\description{
Get buildings data
Expects assessment.dbname, assessment.host, assessment.user, assessment.password to be set for access to the assessment database
}
