% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{get_tax}
\alias{get_tax}
\title{Get tax data
Expects assessment.dbname, assessment.host, assessment.user, assessment.password to be set for access to the assessment database}
\usage{
get_tax(year = 2017, tax_coord = NA)
}
\description{
Get tax data
Expects assessment.dbname, assessment.host, assessment.user, assessment.password to be set for access to the assessment database
}
