% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\docType{data}
\name{land_use_labels}
\alias{land_use_labels}
\title{Land Use Labels}
\format{An object of class \code{list} of length 25.}
\usage{
land_use_labels
}
\description{
Land Use Labels
}
\keyword{datasets}
