% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{residential_land_use}
\alias{residential_land_use}
\title{Filter to just get properties with residential land use}
\usage{
residential_land_use(data)
}
\description{
Filter to just get properties with residential land use
}
