% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\docType{data}
\name{residential_land_use_codes}
\alias{residential_land_use_codes}
\title{Residential land use codes}
\format{An object of class \code{character} of length 7.}
\usage{
residential_land_use_codes
}
\description{
Residential land use codes
}
\keyword{datasets}
